Refactor quiz starting
Quiz starting now uses a chained series of http requests that first creates the quiz submission and then refetches all the quiz submissions using the quiz_submissions_info endpoint, so that we have closed_at and is_open information for individual quiz submissions. In this way we can more easily make the UI reflect the state of the data offloading the work to the database instead of the client.