Add reset_db script and run it before test_db. Closes #114
I often see the database get in a weird state such that tests fail. I don't know how it gets in that state. It appears to be some weird rollback issue that effects the /test/yeluke-quiz-answer.sq script. Well, rather than get to the bottom of that I'm going to just reset the database before running the tests every time. This shell script does that and the shell script is run when you do `npm run test_db`