Make it possible to update assignments using list of files
Previously all the assignments has to be in a single file. Now
you can split them up into individual assignment files, each of
which is a YAML dictionary. Also it is now OK to have keys that
do not appear in the database in those dictionaries. They will
not affect the insert.